src/apps/models/speech_recognition_model/requirements.txt
speechrecognition
python-dotenv
pyttsx3
openai
langchain
pygame
pyaudio